refactor: move gripper function to gripper.c
