+++ gripper_17.X/include/system_init.h
@@ -15,24 +15,15 @@
 #include "rtc.h"
 #include "sercom1_i2c.h"  // Encoder i2c
 #include "tc0.h"
+#include "tc1.h"
 #include "tcc.h"
 #include "tcc0.h"
 #include "tcc_common.h"
-#include "tc1.h"
 #include "wdt.h"
 
 #ifdef DEBUG
 #include "usart.h"
 #endif
-
-// Encoder
-#define SHOULDER_ADDR 0x40
-#define WRIST_ADDR 0x41
-#define GRIP_ADDR 0x42
-#define ANGLE_REGISTER 0xFE
-#define I2C_TIMEOUT 100000
-#define NUM_ENCODERS 3
-
 
 #define CAN_SEND_ANGLES 0x469
 
@@ -40,14 +31,11 @@
 #define ADC_VREF 5.0f
 #define CURRENT_TRESHOLD 2.7f  // 1 A
 #define VOLTAGE_THRESHOLD 2048
-#define RTC_COMPARE_VAL 50
-
 
 #define EVENT_SET_PWM (1 << 3)
 #define EVENT_READ_ENCODER (1 << 4)
 #define EVENT_START_GRIPPER (1 << 5)
 #define EVENT_TRANSMIT_ANGLES (1 << 3)
-
 
 #ifdef __cplusplus
 extern "C" {
@@ -65,7 +53,6 @@ typedef enum {
     RESET_MCU,
 } CAN_RX_ID;
 
-
 typedef enum {
     SERVO_1,
     SERVO_2,
@@ -77,7 +64,6 @@ struct can_tx_frame {
     uint8_t buf[64];
     uint8_t len;
 };
-
 
 struct can_rx_frame {
     uint32_t id;
@@ -92,23 +78,6 @@ struct can_rx_frame {
  */
 void system_init(void);
 
-static inline void stop_gripper(void) {
-    WDT_Disable();
-    PORT_REGS->GROUP[0].PORT_OUTCLR = (1 << 0) | (1 << 27) | (1 << 28);
-    TCC0_PWMStop();
-    TCC1_PWMStop();
-    ADC0_Disable();
-}
-
-static inline void start_gripper(void) {
-    TCC0_PWMStart();
-    TCC1_PWMStart();
-    ADC0_Enable();
-    RTC_Timer32Start();
-    RTC_Timer32CompareSet(RTC_COMPARE_VAL);
-    PORT_REGS->GROUP[0].PORT_OUTSET = (1 << 0) | (1 << 27) | (1 << 28);
-}
-
 static inline bool can_transmit(struct can_tx_frame* frame) {
     return CAN0_MessageTransmit(frame->id, frame->len, frame->buf,
                                 CAN_MODE_FD_WITHOUT_BRS,
@@ -116,8 +85,9 @@ static inline bool can_transmit(struct can_tx_frame* frame) {
 }
 
 static inline bool can_recieve(struct can_rx_frame* frame) {
-    return CAN0_MessageReceive(&frame->id, &frame->len, frame->buf, &frame->timestamp,
-                        CAN_MSG_ATTR_RX_FIFO0, &frame->msg_atr);
+    return CAN0_MessageReceive(&frame->id, &frame->len, frame->buf,
+                               &frame->timestamp, CAN_MSG_ATTR_RX_FIFO0,
+                               &frame->msg_atr);
 }
 
 #ifdef __cplusplus
