+++ gripper_17.X/nbproject/configurations.xml
@@ -18,6 +18,7 @@
       <itemPath>include/usart.h</itemPath>
       <itemPath>include/wdt.h</itemPath>
       <itemPath>include/pm.h</itemPath>
+      <itemPath>include/gripper.h</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
@@ -49,6 +50,7 @@
       <itemPath>src/wdt.c</itemPath>
       <itemPath>src/tc0.c</itemPath>
       <itemPath>src/tc1.c</itemPath>
+      <itemPath>src/gripper.c</itemPath>
     </logicalFolder>
   </logicalFolder>
   <sourceRootList>
@@ -134,6 +136,8 @@
         <property key="memories.programmemory" value="true"/>
         <property key="memories.programmemory.ranges" value="0-1ffff"/>
         <property key="poweroptions.powerenable" value="false"/>
+        <property key="programmerToGoFilePath"
+                  value="/home/nathaniel/vortex-auv-embedded/gripper_17.X/debug/default/gripper_17_ptg"/>
         <property key="programoptions.eraseb4program" value="true"/>
         <property key="programoptions.preservedataflash" value="false"/>
         <property key="programoptions.preservedataflash.ranges"
